feat: add subtract function
