+++ package/__init__.py
@@ -1,23 +1,7 @@
 """Package."""
 
+from .math import add, subtract
+
 __version__ = "0.1.0"
 
-
-def add(a: int, b: int) -> int:
-    """
-    Add two numbers.
-
-    Parameters
-    ----------
-    a : int
-        First number.
-    b : int
-        Second number.
-
-    Returns
-    -------
-    int
-        Sum of the two numbers.
-
-    """
-    return a + b
+__all__ = ["add", "subtract"]
