+++ package/math.py
@@ -1,0 +1,41 @@
+"""Math functions."""
+
+
+def add(a: int, b: int) -> int:
+    """
+    Add two numbers.
+
+    Parameters
+    ----------
+    a : int
+        First number.
+    b : int
+        Second number.
+
+    Returns
+    -------
+    int
+        Sum of the two numbers.
+
+    """
+    return a + b
+
+
+def subtract(a: int, b: int) -> int:
+    """
+    Subtract two numbers.
+
+    Parameters
+    ----------
+    a : int
+        First number.
+    b : int
+        Second number.
+
+    Returns
+    -------
+    int
+        Difference of the two numbers.
+
+    """
+    return a - b
